Put GetMiddleware() function into its own separate file
